refactor: extract shared stretch types from composite and mosaic wizards
Consolidates duplicated StretchMethod type, STRETCH_OPTIONS array, and
base stretch params into a new StretchTypes.ts. Removes MosaicStretchMethod
and MOSAIC_STRETCH_OPTIONS — both wizards now import from the shared module.
Also aligns mosaic default asinhA from 0.1 to 0.05 to match composite.
